refactoring: bank_balance becoming basic-api-fetcher
  adding stream for fetching additional token data
