--- source-bank-balance-fetcher/sample_files/config-example.json
@@ -1,4 +1,0 @@
-{
-  "url": "https://config.test.bi.status.im/api/bank",
-  "api_key": "something"
-}
