--- source-bank-balance-fetcher/source_bank_balance_fetcher/__init__.py
@@ -1,8 +1,0 @@
-#
-# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
-#
-
-
-from .source import SourceBankBalanceFetcher
-
-__all__ = ["SourceBankBalanceFetcher"]
