+++ source-basic-api-fetcher/main.py
@@ -2,7 +2,7 @@
 # Copyright (c) 2023 Airbyte, Inc., all rights reserved.
 #
 
-from source_bank_balance_fetcher.run import run
+from source_basic_api_fetcher.run import run
 
 if __name__ == "__main__":
     run()
