+++ source-basic-api-fetcher/metadata.yaml
@@ -10,16 +10,16 @@ data:
   connectorSubtype: api
   connectorType: source
   definitionId: 1c448bfb-8950-478c-9ae0-f03aaaf4e920
-  dockerImageTag: 0.0.2
-  dockerRepository: harbor.status.im/status-im/airbyte/source-bank-balance-fetcher
-  githubIssueLabel: source-bank-balance-fetcher
-  icon: bank-balance-fetcher.svg
+  dockerImageTag: 0.0.1
+  dockerRepository: harbor.status.im/status-im/airbyte/source-basic-api-fetcher
+  githubIssueLabel: source-basic-api-fetcher
+  icon: basic-api-fetcher.svg
   license: MIT
-  name: Bank Balance Fetcher
+  name: Basic API Fetcher
   releaseDate: TODO
   supportLevel: community
   releaseStage: alpha
-  documentationUrl: https://docs.airbyte.com/integrations/sources/bank-balance-fetcher
+  documentationUrl: https://docs.airbyte.com/integrations/sources/basic-api-fetcher
   tags:
     - language:python
 metadataSpecVersion: "1.0"
