+++ source-basic-api-fetcher/source_basic_api_fetcher/__init__.py
@@ -1,0 +1,8 @@
+#
+# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
+#
+
+
+from .source import SourceBasicApiFetcher
+
+__all__ = ["SourceBasicApiFetcher"]
