+++ source-basic-api-fetcher/source_basic_api_fetcher/schemas/bank_balance.json
@@ -1,0 +1,67 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "type": "object",
+  "properties": {
+    "account_id": {
+      "type": [
+        "null",
+        "string"
+      ]
+    },
+    "bank_name": {
+      "type": [
+        "null",
+        "string"
+      ]
+    },
+    "ccy": {
+      "type": [
+        "null",
+        "string"
+      ]
+    },
+    "tier": {
+      "type": [
+        "null",
+        "string"
+      ]
+    },
+    "volatility": {
+      "type": [
+        "null",
+        "string"
+      ]
+    },
+    "asset_type": {
+      "type": [
+        "null",
+        "string"
+      ]
+    },
+    "account_owner": {
+      "type": [
+        "null",
+        "string"
+      ]
+    },
+    "balance_date": {
+      "type": [
+        "null",
+        "string"
+      ]
+    },
+    "balance": {
+      "type": [
+        "null",
+        "number"
+      ]
+    },
+    "balance_usd": {
+      "type": [
+        "null",
+        "number"
+      ]
+    }
+
+  }
+}
